tidy error handling and logging
