+++ src/lib/app-routes/common/auth/keys.js
@@ -41,7 +41,6 @@ exports.register = function(server, options, next) {
     const redisClient =  server.plugins['hapi-redis'].client;
     const authUtils = server.plugins.utilities.auth;
     const invalidCookie = casCookieUtils.invalidate();
-    const errorLogger = server.plugins.logUtil;
 
     const LoginRecord = server.plugins.bookshelf.model('LoginRecord');
 
@@ -121,8 +120,7 @@ exports.register = function(server, options, next) {
                 /**
                  * log error and inserts record to db for login failure
                  */
-                const logError = (err)=> {
-                    errorLogger.logError(['login'], err);
+                const handleError = (err)=> {
                     recordObj.success = 0;
                     saveRecordObj(recordObj);
                     reply(boom.wrap(err));
@@ -155,7 +153,7 @@ exports.register = function(server, options, next) {
                     .then(serveHawkCredentials)
                     .then(_.noop)
                     .then(_.partial(saveRecordObj, recordObj))
-                    .catch(logError);
+                    .catch(handleError);
             }
         }
     });
