+++ src/lib/utils/authentication.js
@@ -108,10 +108,7 @@ module.exports.register = function(server, options, next) {
             .where({ username: encryptedUsername })
             .fetchAll({ withRelated: 'role' })
             .then(resetStudyRoles)
-            .then(getStudyRoles)
-            .catch((err) => {
-                errorLogger.logAndThrowError(['refreshRoles'], err);
-            });
+            .then(getStudyRoles);
     }
 
     /**
@@ -136,8 +133,6 @@ module.exports.register = function(server, options, next) {
                 });
 
                 return Bluebird.all(revokes);
-            }).catch((err) => {
-                errorLogger.logAndThrowError(['revokeRoles'], err);
             });
 
     }
@@ -159,10 +154,7 @@ module.exports.register = function(server, options, next) {
 
         });
 
-        return Bluebird.all(rolePromises)
-            .catch((err) => {
-                errorLogger.logAndThrowError(['addRoles'], err);
-            });
+        return Bluebird.all(rolePromises);
     }
 
     /**
@@ -273,10 +265,7 @@ module.exports.register = function(server, options, next) {
             .fetch()
             .then(checkAccountStatus)
             .then(comparePassword)
-            .then(handleCompare)
-            .catch(function(err) {
-                errorLogger.logAndThrowError(['validate-user'], err);
-            });
+            .then(handleCompare);
     };
 
     server.expose('auth', authUtils);
