+++ src/lib/utils/register-unhandled-rejections.js
@@ -1,0 +1,13 @@
+'use strict';
+const Bluebird = require('bluebird');
+Bluebird.longStackTraces();
+module.exports.register = (server, options, next) => {
+    process.on(
+        'unhandledRejection',
+        server.plugins.globalErrorHandler.handleUncaught
+    );
+    next();
+};
+module.exports.register.attributes = {
+    name:'registerUnhandledRejection'
+};
